[zorg] Change static analyzer benchmark jenkins bot to be non-strict.
For now, we'll just check to make sure there are no failures/crashes rather
than that the analysis results exactly match the expected results.

The analysis results are heavily dependent on headers, so this will make it
easier to reproduce and troubleshoot bot failures without needing to
exactly recreate SDK/headers on the bots.