LLDB Xcode build: run python script directly
The Green Dragon LLDB macOS bot has been failing to run the test
suite on x86_64 due to the Xcode target permuting the compiler
environment in a way that foiled finding SDKs.  This likely was
introduced via changes in the newer Xcode that replaced the old
bot.

This change switches over to parsing the build settings from
the xcodebuild command line, and directly calling the Python-based
LLDB test suite.  This avoids the issue where clang's environment
was getting mutated.  (The Xcode target approach previously
existed solely to isolate the zorg build script from having to
know anything about the Xcode build configuration, but we now
handle that another way.)