Build clang and LLD earlier in the Windows asan bot config
Right now the "check" step builds LLVM, Clang, and LLD, which is an
awful lot of compilation to do during a testing step. Doing that work up
front will make it obvious if there's a compilation issue in Clang or
LLD without having to click through to the testing logs.